Move schain_allocation into tests
