+++ tests/schain_allocation.py
